+++ DatadogSessionReplay/Sources/Recorder/ViewTreeSnapshotProducer/ViewTreeSnapshot/NodeRecorders/UITabBarRecorder.swift
@@ -19,18 +19,16 @@ internal final class UITabBarRecorder: NodeRecorder {
             wireframeRect: inferOccupiedFrame(of: tabBar, in: context),
             wireframeID: context.ids.nodeID(view: tabBar, nodeRecorder: self),
             attributes: attributes,
-            color: inferColor(of: tabBar)
+            color: inferBackgroundColor(of: tabBar)
         )
 
         let node = Node(viewAttributes: attributes, wireframesBuilder: builder)
-        let subtreeRecordingResults = recordSubtree(of: tabBar, in: context)
-        let allNodes = [node] + subtreeRecordingResults.nodes
-        let resources = subtreeRecordingResults.resources
+        let allNodes = [node] + recordSubtree(of: tabBar, in: context)
 
-        return SpecificElement(subtreeStrategy: .ignore, nodes: allNodes, resources: resources)
+        return SpecificElement(subtreeStrategy: .ignore, nodes: allNodes)
     }
 
-    private func recordSubtree(of tabBar: UITabBar, in context: ViewTreeRecordingContext) -> RecordingResult {
+    private func recordSubtree(of tabBar: UITabBar, in context: ViewTreeRecordingContext) -> [Node] {
         let subtreeViewRecorder = ViewTreeRecorder(
             nodeRecorders: [
                 UIImageViewRecorder(
@@ -78,7 +76,7 @@ internal final class UITabBarRecorder: NodeRecorder {
         return occupiedFrame
     }
 
-    private func inferColor(of tabBar: UITabBar) -> CGColor {
+    private func inferBackgroundColor(of tabBar: UITabBar) -> CGColor {
         if let color = tabBar.backgroundColor {
             return color.cgColor
         }