+++ DatadogSessionReplay/Sources/Recorder/ViewTreeSnapshotProducer/ViewTreeSnapshot/ViewTreeRecorder.swift
@@ -7,11 +7,6 @@
 #if os(iOS)
 import UIKit
 
-internal struct RecordingResult {
-    let nodes: [Node]
-    let resources: [Resource]
-}
-
 internal struct ViewTreeRecorder {
     /// An array of enabled node recorders.
     ///
@@ -20,18 +15,16 @@ internal struct ViewTreeRecorder {
     let nodeRecorders: [NodeRecorder]
 
     /// Creates `Nodes` for given view and its subtree hierarchy.
-    func record(_ anyView: UIView, in context: ViewTreeRecordingContext) -> RecordingResult {
+    func record(_ anyView: UIView, in context: ViewTreeRecordingContext) -> [Node] {
         var nodes: [Node] = []
-        var resources: [Resource] = []
-        recordRecursively(nodes: &nodes, resources: &resources, view: anyView, context: context)
-        return RecordingResult(nodes: nodes, resources: resources)
+        recordRecursively(nodes: &nodes, view: anyView, context: context)
+        return nodes
     }
 
     // MARK: - Private
 
     private func recordRecursively(
         nodes: inout [Node],
-        resources: inout [Resource],
         view: UIView,
         context: ViewTreeRecordingContext
     ) {
@@ -48,14 +41,11 @@ internal struct ViewTreeRecorder {
         if !semantics.nodes.isEmpty {
             nodes.append(contentsOf: semantics.nodes)
         }
-        if !semantics.resources.isEmpty {
-            resources.append(contentsOf: semantics.resources)
-        }
 
         switch semantics.subtreeStrategy {
         case .record:
             for subview in view.subviews {
-                recordRecursively(nodes: &nodes, resources: &resources, view: subview, context: context)
+                recordRecursively(nodes: &nodes, view: subview, context: context)
             }
         case .ignore:
             break